Fixed checkstyle problems.
