+++ src/main/java/ru/planetnails/partnerslk/model/order/CANCELED.java
@@ -1,33 +1,32 @@
 package ru.planetnails.partnerslk.model.order;
 
 import lombok.extern.slf4j.Slf4j;
-import org.slf4j.Logger;
+import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Component;
 import ru.planetnails.partnerslk.exception.NotFoundException;
 import ru.planetnails.partnerslk.model.order.dto.OrderMapper;
 import ru.planetnails.partnerslk.model.order.dto.OrderOutDto;
 import ru.planetnails.partnerslk.repository.OrderRepository;
 
-import javax.validation.ValidationException;
 import java.time.LocalDateTime;
 
 @Slf4j
-@Component("APPROVED")
-public class putStatusApprovedForOrder implements OrderGenerator {
+@Component("CANCELED")
+public class CANCELED implements OrderGenerator {
 
     private final OrderRepository orderRepository;
 
-    public putStatusApprovedForOrder(OrderRepository orderRepository) {
+    @Autowired
+    public CANCELED(OrderRepository orderRepository) {
         this.orderRepository = orderRepository;
     }
 
-
     @Override
     public OrderOutDto setStatusForOrderUser(String orderId, String user) {
-        Order order = validation(orderId, user, orderRepository, log);
-        order.setStatus(OrderStatus.APPROVED);
+        Order order = APPROVED.validation(orderId, user, orderRepository, log);
+        order.setStatus(OrderStatus.CANCELED);
         VtOrderStatuses vtOrderStatuses = new VtOrderStatuses(
-                OrderStatus.APPROVED,
+                OrderStatus.CANCELED,
                 LocalDateTime.now(),
                 user
         );
@@ -42,26 +41,15 @@ public class putStatusApprovedForOrder implements OrderGenerator {
         Order order = orderRepository.findById(orderId)
                 .orElseThrow(() -> new NotFoundException("Order not found"));
         log.info("Order with id {} found", orderId);
-        order.setStatus(OrderStatus.APPROVED);
+        order.setStatus(OrderStatus.CANCELED);
         VtOrderStatuses vtOrderStatuses = new VtOrderStatuses(
-                OrderStatus.APPROVED,
+                OrderStatus.CANCELED,
                 LocalDateTime.now(),
                 user
         );
         vtOrderStatuses.setOrder(order);
         order.getVtOrderStatuses().add(vtOrderStatuses);
 
-        return order;
-    }
-
-    protected static Order validation(String orderId, String user, OrderRepository orderRepository, Logger log) {
-        Order order = orderRepository.findById(orderId)
-                .orElseThrow(() -> new NotFoundException("Order not found"));
-        log.info("Order with id {} found", orderId);
-        if (!user.equals(order.getVtOrderStatuses().get(0).getUser())) {
-            log.error("Order doesn't belong to User with id {}", user);
-            throw new ValidationException("Order doesn't belong to this User");
-        }
         return order;
     }
 }
