+++ src/main/java/ru/planetnails/partnerslk/model/order/COMPLETED.java
@@ -4,20 +4,36 @@ import lombok.extern.slf4j.Slf4j;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Component;
 import ru.planetnails.partnerslk.exception.NotFoundException;
+import ru.planetnails.partnerslk.model.order.dto.OrderMapper;
 import ru.planetnails.partnerslk.model.order.dto.OrderOutDto;
 import ru.planetnails.partnerslk.repository.OrderRepository;
 
 import java.time.LocalDateTime;
 
 @Slf4j
-@Component("WORKHOUSE")
-public class putStatusWorkhouseForOrderManager implements OrderGenerator {
+@Component("COMPLETED")
+public class COMPLETED implements OrderGenerator {
 
     private final OrderRepository orderRepository;
 
     @Autowired
-    public putStatusWorkhouseForOrderManager(OrderRepository orderRepository) {
+    public COMPLETED(OrderRepository orderRepository) {
         this.orderRepository = orderRepository;
+    }
+
+    @Override
+    public OrderOutDto setStatusForOrderUser(String orderId, String user) {
+        Order order = APPROVED.validation(orderId, user, orderRepository, log);
+        order.setStatus(OrderStatus.COMPLETED);
+        VtOrderStatuses vtOrderStatuses = new VtOrderStatuses(
+                OrderStatus.COMPLETED,
+                LocalDateTime.now(),
+                user
+        );
+        vtOrderStatuses.setOrder(order);
+        order.getVtOrderStatuses().add(vtOrderStatuses);
+
+        return OrderMapper.fromOrderToOrderOutDto(orderRepository.save(order));
     }
 
     @Override
@@ -25,9 +41,9 @@ public class putStatusWorkhouseForOrderManager implements OrderGenerator {
         Order order = orderRepository.findById(orderId)
                 .orElseThrow(() -> new NotFoundException("Order not found"));
         log.info("Order with id {} found", orderId);
-        order.setStatus(OrderStatus.WORKHOUSE);
+        order.setStatus(OrderStatus.COMPLETED);
         VtOrderStatuses vtOrderStatuses = new VtOrderStatuses(
-                OrderStatus.WORKHOUSE,
+                OrderStatus.COMPLETED,
                 LocalDateTime.now(),
                 user
         );
@@ -35,10 +51,5 @@ public class putStatusWorkhouseForOrderManager implements OrderGenerator {
         order.getVtOrderStatuses().add(vtOrderStatuses);
 
         return order;
-    }
-
-    @Override
-    public OrderOutDto setStatusForOrderUser(String orderId, String user) {
-        return null;
     }
 }
