+++ src/main/java/ru/planetnails/partnerslk/model/order/WORKHOUSE.java
@@ -4,35 +4,20 @@ import lombok.extern.slf4j.Slf4j;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Component;
 import ru.planetnails.partnerslk.exception.NotFoundException;
-import ru.planetnails.partnerslk.model.order.dto.OrderMapper;
 import ru.planetnails.partnerslk.model.order.dto.OrderOutDto;
 import ru.planetnails.partnerslk.repository.OrderRepository;
 
 import java.time.LocalDateTime;
 
 @Slf4j
-@Component("CANCELED")
-public class putStatusCanceledForOrder implements OrderGenerator {
+@Component("WORKHOUSE")
+public class WORKHOUSE implements OrderGenerator {
 
     private final OrderRepository orderRepository;
+
     @Autowired
-    public putStatusCanceledForOrder(OrderRepository orderRepository) {
+    public WORKHOUSE(OrderRepository orderRepository) {
         this.orderRepository = orderRepository;
-    }
-
-    @Override
-    public OrderOutDto setStatusForOrderUser(String orderId, String user) {
-        Order order = putStatusApprovedForOrder.validation(orderId, user, orderRepository, log);
-        order.setStatus(OrderStatus.CANCELED);
-        VtOrderStatuses vtOrderStatuses = new VtOrderStatuses(
-                OrderStatus.CANCELED,
-                LocalDateTime.now(),
-                user
-        );
-        vtOrderStatuses.setOrder(order);
-        order.getVtOrderStatuses().add(vtOrderStatuses);
-
-        return OrderMapper.fromOrderToOrderOutDto(orderRepository.save(order));
     }
 
     @Override
@@ -40,9 +25,9 @@ public class putStatusCanceledForOrder implements OrderGenerator {
         Order order = orderRepository.findById(orderId)
                 .orElseThrow(() -> new NotFoundException("Order not found"));
         log.info("Order with id {} found", orderId);
-        order.setStatus(OrderStatus.CANCELED);
+        order.setStatus(OrderStatus.WORKHOUSE);
         VtOrderStatuses vtOrderStatuses = new VtOrderStatuses(
-                OrderStatus.CANCELED,
+                OrderStatus.WORKHOUSE,
                 LocalDateTime.now(),
                 user
         );
@@ -50,5 +35,10 @@ public class putStatusCanceledForOrder implements OrderGenerator {
         order.getVtOrderStatuses().add(vtOrderStatuses);
 
         return order;
+    }
+
+    @Override
+    public OrderOutDto setStatusForOrderUser(String orderId, String user) {
+        return null;
     }
 }
